remember to hide this repo zain

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
-user_monthly_income = (int)(input("Enter your monthly income: "))
-user_monthly_expenses = (int)(input("Enter your total monthly expenses: "))
+user_monthly_income = (float)(input("Enter your monthly income: "))
+user_monthly_expenses = (float)(input("Enter your total monthly expenses: "))
 monthly_savings = user_monthly_income  -  user_monthly_expenses
-projected_savings = monthly_savings * 12  + (monthly_savings * 12  * 0.05)
+projected_savings = (float)(monthly_savings * 12  + (monthly_savings * 12  * 0.05))
 print(f"Your monthly savings are ${monthly_savings}.")
 print(f"Projected savings after one year, with interest, is ${projected_savings}.")
